requirements.txt
pylint==1.7.1
azure-storage==0.34.3
azure-batch==3.0.0
xunitmerge==1.0.4
requests